+++ djangoevents/tests/test_schema.py
@@ -79,7 +79,7 @@ def test_validate_valid_event():
     assert ret is True
 
 
-@override_settings(BASE_DIR='/path/to/proj/')
+@override_settings(BASE_DIR='/path/to/proj/src/')
 def test_valid_event_to_schema_path():
     from .test_domain import SampleEntity
